Serialize custom byte block vector

+++ lib/fst/interface/fststore.cpp
@@ -42,6 +42,7 @@
 #include <logical/logical_v10.h>
 #include <integer64/integer64_v11.h>
 #include <byte/byte_v12.h>
+#include <byteblock/byteblock_v13.h>
 
 #include <xxhash.h>
 #include "byteblock/byteblock_v13.h"
@@ -330,7 +331,7 @@ void FstStore::fstWrite(IFstTable &fstTable, const int compress) const
   *p_primChunksetIndex     = 0;
   *p_secChunksetIndex      = 0;
 
-  const unsigned long long nrOfRows = fstTable.NrOfRows();
+  const uint64_t nrOfRows = fstTable.NrOfRows();
   *p_nrOfRows                 = nrOfRows;
   *p_nrOfChunksetCols         = nrOfCols;
   *p_freeBytes4               = 0;
@@ -503,7 +504,7 @@ void FstStore::fstWrite(IFstTable &fstTable, const int compress) const
     {
         colTypes[colNr] = 13;
         IByteBlockColumn* p_byte_block = fstTable.GetByteBlockWriter(colNr);
-        fdsWriteByteBlockVec_v13(myfile, p_byte_block, nrOfRows, compress);
+        fdsWriteByteBlockVec_v13(myfile, p_byte_block, nrOfRows, static_cast<uint32_t>(compress));
         break;
     }
 
@@ -849,7 +850,7 @@ void FstStore::fstRead(IFstTable &tableReader, IStringArray* columnSelection, co
 
   // Check range of selected rows
   const long long firstRow = startRow - 1;
-  const unsigned long long nrOfRows = *p_chunkRows;  // TODO: check for row numbers > INT_MAX !!!
+  const uint64_t nrOfRows = *p_chunkRows;  // TODO: check for row numbers > INT_MAX !!!
 
   if (nrOfRows != 0 && (firstRow >= static_cast<long long>(nrOfRows) || firstRow < 0))
   {
@@ -889,7 +890,7 @@ void FstStore::fstRead(IFstTable &tableReader, IStringArray* columnSelection, co
       throw(runtime_error("Column selection is out of range."));
     }
 
-    const unsigned long long pos = blockPos[colNr];
+    const uint64_t pos = blockPos[colNr];
     const short int scale = colScales[colNr];
 
     switch (colTypes[colNr])
@@ -976,7 +977,7 @@ void FstStore::fstRead(IFstTable &tableReader, IStringArray* columnSelection, co
       std::unique_ptr<IInt64Column> int64ColumP(columnFactory->CreateInt64Column(length, static_cast<FstColumnAttribute>(colAttributeTypes[colNr]), scale));
       IInt64Column* int64Column = int64ColumP.get();
       tableReader.SetInt64Column(int64Column, colSel);
-      fdsReadInt64Vec_v11(myfile, int64Column->Data(), pos, firstRow, length, nrOfRows);
+      fdsReadInt64Vec_v11(myfile, int64Column->Data(), pos, firstRow, static_cast<uint64_t>(length), nrOfRows);
       break;
 	  }
 
@@ -994,6 +995,7 @@ void FstStore::fstRead(IFstTable &tableReader, IStringArray* columnSelection, co
     case 13:
     {
       IByteBlockColumn* byte_block = tableReader.add_byte_block_column(colSel);
+
       read_byte_block_vec_v13(myfile, byte_block, pos, firstRow, length, nrOfRows);
       break;
     }
